fix 2-execute_a_command.pp

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,5 +1,6 @@
-# Execute a command
+#Execute a command
 
 exec {'killmenow':
   command => 'pkill -f killmenow',
+  onlyif => 'pgrep -f killmenow',
 }
